Update main.c
since it is a generic stack , the stack provider does not know what are the elements that are being pushed into the stack like wise the stack user does not know in what way the stack is being implemented? whether it is being implemented using linked list or array.so it usually becomes difficult to write the display function. but acting as an intermediate we can write 3 separate display functions by passing 3 struct stack i.e (3 struct stack variables). then make sure to use %p as the format specifier for pointers in the display function.